Move api/items routes out of app.js
Before this, we were doing all the routing for the
api/items path in our app.js file. I created a routes
folder, with a subdirectory of api and added an item.js
folder which now uses express Router to handles all routes
to api/items. App.js requires this new doc and app.use's it