Extract database configuration from app.js to database.js
Before this, all the logic for deciding which database to
connect to was stored in the app.js file. I ahve extracted
this logic and required the doc in the app.js folder. After
this point, you can require the database.js file and use
the 'databaseUrl' to decide which database to connect to